Retrieve display name from UAA userinfo endpoint
The JSON Web Tokens returned by UAA contain only
user name and email address, but not the user's display name.
For that an additional request to the UAA /userinfo endpoint
is necessary. Note that the UAA client configured for
Gerrit must have the scope "openid" for this to work
as described in the configuration documentation.

Change-Id: I5d1169da88cb32c8eb4133e9710df7a200d7d74a
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>