Set the auth.logouturl property during init
When user signed out in the UI only the web session in Gerrit
was invalidated, but not that in UAA. Consequently when the
user clicked on "Sign In" again the UAA did not present the
login screen but directly returned a new access token. This
behavior can be prevented by configuring an explicit logout
page with the auth.logouturl in the Gerrit configuration.

This patch enhances the plugin's init step to configure the
auth.logouturl property to point to /logout.do in UAA. This
will close the UAA web session and the user can login with
different credentials again.

Change-Id: Id30b9d0d0836439040c05867a0d82e14333d8563
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>