Respect auth.userNameToLowerCase
The CloudFoundry UAA in general is preserving the case
of user names, if attached to another user store, e.g.
an LDAP system or database. Therefore, the plugin should
respect the auth.userNameToLowerCase configuration
parameter. If this parameter is set to true, the plugin
will return the username and externalId parameters in
OAuthUserInfo in lowercase, but it will not touch the
access tokens since the latter have a signature.
However, when comparing user names with the "user_name"
or "sub" attributes of access tokens the correct case
conversion will be applied.

Change-Id: I9198809ab3d52d9a05ad96082cf2667981042bd9
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>