Add redirect parameter to logout URL
The logout page of UAA has the capability to redirect back to an
arbitrary URL [1]. Change the init step to write a logout URL
with a redirect parameter pointing to the canonical web URL of
Gerrit, so that a user can sign in again immediately.

Note, recent versions of UAA switch off the redirect by default
for security reasons. Make sure to uncomment the "logout" section
in the login.yml configuration file of the UAA and add the
Gerrit canonical web URL to the whitelisted redirect URLs.

[1] https://github.com/cloudfoundry/uaa/blob/master/docs/login/Login-APIs.md#logout-get-logoutdo

Change-Id: I0f581ff6897e77ffc5ec5677ab39a366b1685d2c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>