Deliver expiresAt and providerId for access token
The REST API for OAuth access tokens requires tokens to
have an expiration date and the identifier of the OAuth
provider. This patch adds the missing attributes.

A constant is introduced in OAuthModule to denote the
export name of the provider. This constant is used
to construct the providerId for the representation
of access tokens in the REST API.

Change-Id: I7334949ff18ede75f33eda55a7087655e7b88aa5
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>