Load TestUtils.java as a library instead of a test source file
The tests of the cfoauth-plugin failed on master branch with 3 failures.
One failure was caused by the TestUtils-class, which did not contain
any method annotated with `@Test`, causing Junit to raise
`java.lang.Exception: No runnable methods tests`.

This change creates a library containing the TestUtils-class, which is
loaded as a test-dependency and excludes TestUtils.java from the
test-sources.

This fixes one of the three observed test failures. The other test
failures will be addressed in a dedicated change.

Change-Id: I393969b46c39757d8bd4c84cda8d3b8a884b7510