Fix assembly of external id
Two tests for the master-branch of the cfoauth-plugin failed:

- testParseAccessTokenResponse
- testGetAsAccessToken

Both tests showed the same ComparisonFailure:

  expected:<external[:]marissa> but was:<external[]marissa>

Since Commit 744d2b89 in Gerrit the scheme for external auth is defined
without the trailing ':'. The colon however was expected by the cfoauth
plugin, causing false assembly of the external id.

This change adds the addition of a colon, when defining the external ID
in the plugin code.

Change-Id: If3d0cc52c0e64fb2b9cf3ce0a4b0774e9533da56