Fix conversion from AccessToken to OAuthToken
The login via UAA failed with "Server Error", when using the cfoauth
plugin with Gerrit stable-2.16, because OAuthToken-constructor provided
by the oauth extension of Gerrit does now require the 'secret'- and the
'raw'-parameter not to be null.

This change adds the raw response of the UAA server to the AccessToken
object, which is then handed to the OAuth extension point of Gerrit
instead of null.

Since UAA uses OAuth 2.0, no token secret is used. Since the Gerrit
OAuth extension point expects a token secret to be sent, an empty string
will be used instead. This follows the approach used by the scribe-java
library, which is used in the OAuth implementation of Gerrit and this
plugin.

Change-Id: I4e36b2a651951aff1cf08932b2902fa86905976e