Fix conversion from AccessToken to OAuthToken
The login via UAA failed with "Server Error", when using the cfoauth
plugin with Gerrit stable-2.16, because OAuthToken-constructor provided
by the oauth extension of Gerrit does now require the 'secret'- and the
'raw'-parameter not to be null.

TODO: Use the expected values as parameters, the currently used values
for secret and raw are just guesses for now.

Change-Id: I4e36b2a651951aff1cf08932b2902fa86905976e