Update patch set 2
Patch Set 2:

Still confusing change especially VRFY+1 vote, because the tests are failing. It would be much cleaner to state in the commit message, that the tests are failing before and after this change and that the failing tests are fixed in the follow-up change. A different option would be to swap the order of this and follow-up change.

Patch-set: 2
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>